Remove unneeded "startup" calls for newly made dijit widgets that may be slowing down page loading.
See: http://dojotoolkit.org/reference-guide/1.10/dijit/layout.html#id8

Essentially, the only time startup needs to be called in the latest
version of Dojo (1.10.x or probably several versions earlier) is when a
dijit is added to the existing document's DOM visual hierarchy. All
child widgets will have
startup called automatically at that time, and dijits added later will
have startup called by the system when they are added. So, essentially,
it seems the only time startup is needed to be called is for before
opening a dialog for the first time, when adding a ContentPane or such
to the document at the top level, or when adding a widget to a document
at the top level. This was not the case with much earlier versions of
Dojo it seems.

Still not sure what is needed if a dijit's domNode is added to another
dom node, sort of bypassing the normal addChild or placeAt process. If
there are bugs from this change, they might be in such cases.