Remove resizeGridsKludge
Dojo dgrids calculate their header height when they are first created.
If the page they are on is not visible (or maybe in some other
circumstances), the header will be zero height and will be overwritten
by the first added item.

A previous solution was to track every grid added to a page and then
resize them when the page was done building and was displayed. However,
that seemed kludgy.

A new approach seems to work of just resizing the page (or nested form)
after it has been created. That seems simpler and less kludge, and might
also prevent or fix other possible similar resizing issues. So the code
has been changed to call resize and the kludge has been removed.