Remove defensive code for adding to pagePane which is no longer needed
Previously, string ids could be passed around for content panes, and so
could dom nodes, and that caused issues when trying to add a grid table
or story browser. That code is no longer needed now that content panes
are passed around consistently. It is also cluttery and so should be
removed.