Remove dependendency on domain in PanelBuilder
The PanelBuilder could (in theory) be a reuseable component. In any
case, it would be good to be able to test it stand-alone. This change
requires a PanelBuilder instance to be configured with a function for
responding to button presses and an array or function to be used to get
panel specifications from a panel ID.