Improve translate support
Removed leading "#" when constructing translation IDs from strings
immediately passed into translate. Keeping it in IDs for buttons and
such declared in various places as it is a useful indicator for the
intent of the string that will be passed around.

Improved translate function to return defaultText if no translation for
tag without leading hash.