Replace PanelBuilder.updateQuestionsForPageChange with different approach
The old updateQuestionsForPageChange method was created when the
architecture was to have several pages at once. In the new architecture,
there is only one page around, so there is less need for it. That method
would also leak memory because it maintained a list which had items
added to it which refered to GUI widgets and the list was never cleaned
up or removed from. 

The new architecture is using dojo topic subscriptions for topics which
signal via domain function names that the function has changed value. A
subscription is added to such topics when a "function" display item is
created.

The way that text to display was calculated was also changed, supporting
a "{{result}}" tag that can be replaced. If the tag is missign, the
calculated text is appended at the end of the string as before.