Remove unneeded code in FieldSpecificationCollection
Now that fields are under panels and their are not field specifications
that are about panels, there is no need to check when building a model
if a field defines a panel. Also, because the fields in a model are
tracked as panels are added, there is no need to check the modelClass of
a field when iterating over the fields already determined to define a
specific model.