Modify saving code to only save data for the current page in a document with a name related to that page's model
Code written a couple days ago, but wondering about design issues.

Realized could still save and load for entire project, and only check
for changes for the current model for the current page, updating the
rest of that data in the background as a project file changes. However,
that multiplies the data saved and transferred by data for all the other
pages in the project (currently around 57). So, on average, that means
about 30X more data if save and load entire project (if each section was
the same). Some activities might produce a lot more data, like the
catalysis pages for theming or perspectives, and that would also be
multiplied by every change.

It would be good also to ensure the Pointrel code is working well in a
general way, so splitting up the file into seperate documents will be a
good test of it. Or a design challenge to improve it if needed.

Another possibility would be to only store data with triples as they
change. However, then the challenge is what document ID to give the
transaction so the triples can be "deleted" if that document is changed.
Could perhaps use a document ID related to the fields if only storing
one triple per transaction. But seems a bit awkward. The general notion
that you have an object and you make changes to it, indexing (or
publishing?) some of that data as triples seems clearer and more
manageable in some way.

Still unsure about best design, but need to move forward.