Handle questionnaire update in story browser
The questionnaire may be null when the application starts up.
Then it may get set to something eventually.
If the story browser is the page the application starts with,
it needs to handle the null changing to a questionnaire.

Also fix bug with story browser not correctly updating totals for
stories with a filter multiselect choice when list of all stories
changes.