Comment out currently unneeded calculation code in questionAnswer
Now that the date is retrieved from triples from server, the calculation
code is wrong since it references model fields. It may still be useful
for translating select choices, but that is not currently needed. If
that is needed, the commented code may be useful for making that
translation work.