Copying current client and server code from Pointrel20150417 project
This is intended to better support multi-user operations.

The new Pointrel version is also intended to better support WordPress.
This new approach should reduce the amount of data needed to be stored
in the WordPress database. It is also designed to be integrated with the
WordPress heartbeat. The new approach is also designed to integrate
easily with the WordPress AJAX api.