Change model.js to work with new Pointrel version (initial support)
This is an initial approach; it may be revised significantly later.

This supports shared editing of text fields.

It does not seem to be working with grids, which need to be modified
further in any case to have more fine-grained sharing of changes.

The bigg changes to model.js is to use new "project" concept from
project.js. The project is essentially just a Pointrel TripleStore. This
may need to be revised later to have more project fields.

To have the least changes from the old approach for models, a page model
is still created for each page (versus, say, one model for the entire
project). Subscriptions are made to the TripleStore for triples that
would affect each model field. The model fields are watched for changes.
A set of shadow "_saved" values are stored in the model for comparison
when responding to changes from the TripleStore or model updates from
GUI to avoid spurious updates (which could otherwise lead to ringing
cycles of messages in the system or other unneeded additions of
triples). Conflicts where the user would lose data being edited are
logged to the console, but are not yet otherwise handled. Other sorts of
conflicts are not yet reported.