Use one project model for all pages instead of making page models as change pages
This is a return to the original conception from months ago, except that
now project field values are written to a triple store as they are
changed.

Still may need more work to be able to track changes to model inside
arrays.

This removes most functionality from the domain object to put it in the
Project class.