Make progress on getting surveys working again (still not done)
Remove some dependencies on storage.

Remove some now (or about to be) out-of-date comments related to loading
questionnaires, results, and status.

Streamlining approach for survey status to (intended) be resturned when
go to retrieve questionnaire.

Removing use of storage from surveyBuilder so it just does GUI things
and callers are responsibel for storage.