Do not require Passport authentication by default
Survey users probably don't need to be authenticated to get support
fields or to read or post JSON messages. However, this could be switched
back for different setups.

Users still need to have permission to read or post using the Pointrel
authentication though. That is set with the Pointrel admin tool.
Anonymous access must be granted explicitly to projects to read
questionnairs or post survey results.