Fix PointrelClient so it discards messages with callbacks if there is an error
The caller will presumably resend these messages.

This was otherwise causing an issue with the NarraFirma survey when a
survey was resent. The resend was being blocked by a previous message
that failed to be sent which remained stuck in the outgoing message
queue, and the callback for the resent message was never called.