Always call updateData on view in GridWithItemPanel; ignore items which are null
Updating data for every view is unfortunate and inefficent. However,
since updates can come from another client, the list can change at any
time even if this client is not changing it outside of the Grid. There
may be more computationally efficient alternatives like checking the
type of update (whetehr it is about the list) or listening for specific
updates (maybe best, but also requires unsubscribing when the component
is destroyed/unloaded). But, for now, going with the simple and reliable
approach until performance becomes an issue.