Move recommendations_filledin.csv from design to (new folder) WebContent/recommendations
It turns out that the ~20K CSV file expands to ~500K when parsed and
converted to JavaScript. So, it seems more efficient to keep it as CSV
and parse it on the client side. There may also be better use of
duplicate strings for object keys this way. So, moving this file to
where the client can load it.