Improve WordPress plugin to calculate SHA256 for JSON correctly to match JavaScript
Also fixes some other bugs in narrafirma.php, including not correctly
getting count of number of records in journal.

The version of PHP used to test does not support a flag for not escaping
non-ascii JSON characters, so there may still be an issue to resolve
here (see comments in the code).