Change valuePathResolver to default to not allowing direct access to object fields or to a tripleStore
Update story browser to supply a stories function instead of a stories
field in another object.

Do not store clientState or project in PanelBuilder but instead pass in
application object to use as context (not fully finished).

This change is a step towards using model objects instead of direct
access to triples. This, along with the previous two changes, breaks the
application. More refactoring needed.