Do not poll for new messages in PointrelClient if document not visible
This is to reduce network traffic and sever load when application is not
in use. The cost may be a bit of delay loading new messages when the
appicaiton is primary again. The visibility API is only availabel in
later browsers. If it is not available, the applicaiton will continue to
poll always.