Change Project topic from "testing" to "narrafirmaProject"
This change relates to how many projects coudl be stroed in one journal,
or how data is copied from journal to journal. The topic could be or
contain the project name instead of being a fixed string related to the
application. However, a project's data is found from a root object with
the name already and so is distinguishable even if multipel projects
were in the same journal (although you can't filter by a topic specific
to the project). Likewise, the project name could be added to the
surveyResults topic string; however survey results also have the project
name in them already. although again you can't filter by topic for a
specific project. The questionnaire remains undistinguished for a
project except by questionnaire name (which multiple projects might
use). May want to revisit this all at some point as the issue becomes
clearer (at a cost of making older data inaccessible without conversion
if these topics change).