Fixed checkboxes data entry bug.
Fixed bug where answering a survey question with check boxes was setting
the value to a string instead of a dictionary. The offending code was
here:

    value[optionValue] = event.target.checked;
    change(null, optionValue);

See it? The second line should have been

    change(null, value);

A tiny typo! Isn't that always the way.

Included in this fix is a backwards-compatible conversion of any stored
string for such a question to a dictionary with that one entry
preserved. You will have to fix the data for that question, but at least
the first item checked will still be there.