Improvements to story card display and story editing in "Review incoming stories" screen.
In story card display, moved annotation questions to end (so they are
more obviously different from survey questions). Added more tags for
styling of parts of story card display (question names, annotation
questions, etc). Improved formatting on "Print story cards" also.

In panels in "Review incoming stories" screen, made it so the annotation
questions are (a) near the top (for quicker editing) and (b) in a
different color/style, so you can see which ones they are. Moved "Reason
to ignore story" thing to bottom, so it's not in your way, and improved
the text to be more quickly understandable.