Fix for error message caused by empty contingency table.
If the maximum x and y count is zero (because there is no data for the
table), it was showing an error because the scale factor was Infinity
(size divided by zero). Changed it so it now attempts to set things at
zero instead, to stop the error from appearing (and maybe slowing down
the graphing).