Fixed earlier mistake about "click to close" on humane message
Apparently "click to close" does not mean the message stays up until you click it; it just means that you can click to close it before the "timeout" time has passed. Increased the timeout time (from 2 to 5 seconds), and you can click BEFORE the 5 seconds are up to close the message.