Added reporting on number of stories told in session
All places where stories are looked at (Review incoming stories, Spot-check graphs, Explore patterns) now have an extra "question" which is the number of stories told by the participant who told that story. This is done by session for survey and data entry input, and by a "Participant ID" field for CSV input. The number of stories told is a measure of volubility, or how much people have to say, that sometimes creates interesting patterns when compared to what people said (e.g., people who told >2 stories tended to be more upset, etc). This is a generated field; it is not saved but is recalculated whenever data is read to be displayed. Luckily the stories are saved in bundles by session, so it was already there in the system.