Added place to write, copy, paste reminders for next project
The idea is to provide a bit more support for cyclical projects. Added a simple place to write some reminders for the next project, then another place to paste them in after copying them from the previous project. Could have done something more elaborate, but it could have been confusing. This will be simple but useful.