Removed lowest options for story count threshold
Statistical tests can't really be meaningful for fewer than 10 stories, even if you sample every member of the population under consideration. So removed the "1" and "5" options. Also added note to warn about using small numbers (10 and 20). Sample size requirements vary by test and assumptions, but generally "a few dozen" is the lower limit for most tests - which equates to 24-36, so I said that 30 is a reasonable minimum. 20 is pushing it; 10 is at the borderline of meaninglessness.