Added number to slider
Can now see number (0 to 100) on slider. Clicking on the number causes a dialog to pop up so you can change the number to a specific value. (This is more for data entry than anything else.) This could be a problem for accessibility, but since it is redundant functionality (you can use the arrow keys to set the slider values) it's okay. Also added a note in the help system that if you don't want to show these numbers you can set display to "none" for that class in the CSS.