Reference system for observation sections
Observations can be separated into sections marked by headers of the form @header@. Interpretations that reference those sections will cause only those sections to print under that interpretation in the report. Explained this in help system. This is because somebody wanted to have more than one observation per graph, and I didn't want to clutter up the interface (on the most complicated page in the app) with yet another list. This method is a bit cryptic for advanced users but invisible to beginners.