Disallowed some actions on interpretation clustering
Realized that users could (a) change names of interpretations in clustering diagram, unmooring them from the observations they were connected to, and (b) create new interpretations unlinked to any observation. So disabled buttons for creating new interpretations and editing interpretations. Also added warnings to page and help system about how the clustering diagram is a snapshot copy of the interpretations on the previous page, and how you will need to delete and re-add a changed interpretation to use the new text. Would have preferred to make it update to changed text, but the clustering diagram structure makes that difficult, and didn't want to break anything. Usually this will not be an issue, and replacing an item is not that hard.