Changes for new image map, plus removed some buttons
Removed second set of buttons for PNI phases on main dashboard. Now clicking on buttons in imagemap is the only way to get to the stages. This could be a problem for accessibility. On Chrome and Safari the tab key includes the imagemap, so that's okay. On Firefox the tab key doesn't include anything on the page, so that's ... okay too.