Moved text fields for catalysis report to configure page
Because I went to change something and couldn't remember what page it was on. Now all the fields that change the report are on the same page. Also noticed that I forgot to include valuePath statements before so the things applied to all reports, not just the selected one. Oops! Fixed now.