Fixed interpretations getting unlinked due to name changes
The way the clustering diagram worked was that you could rename an interpretation there, and the catalysis report used the diagram to organize the report. However, because the name was used to LINK the interpretation to its observation, renaming an interpretation could cause it to become unlinked. Added a "referenceUUID" field to the clustering item so that the original interpretation could be checked for changes to its name. Also added a check to update this during printing. User still needs to click the Update button for changes made to interpretations, but now it updates names and texts as well as added (new) interpretations.