Sanitized observation, interpretation, perspective texts
Now these should show in the report as valid HTML. While doing this, noticed that exceptions in the sanitize method were not being caught in every case of user supplied information, and user saw nothing but a blank screen. Now catching all exceptions in sanitizing of user text and putting up alert so user has a clue to what happened.