Added ideas to printed catalysis report
Note that this addition uses the new referenceUUID field to find the ideas. This field is being added to clustering diagram items (that are intepretations) when the user clicks the "Start or update clustering diagram" button on the "Cluster interpretations" page. This means that catalyis reports created in a version of NF prior to 1.0.0 will not show printed ideas until the clustering diagram is updated. I don't think very many people will be affected by this little bit of backward incompatibility.