Better error handling for missing items
Added more "displayVisible" things so page elements don't display if no story collection or catalysis report is selected. Dealt better with pathological condition of the user creating a story collection or catalysis report and not giving it a name, which used to cause gobbledgygook to appear in the drop-down list. Now it gives a warning that the thing should have a name. Also added warnings about how collections and reports should have names.