Fixed mistake where navigation bar was scrolling
I have no idea how this happened, but while making screenshots I realized that during the color redesign I changed the position: fixed on the navigation bar to position: absolute, and now the nav bar doesn't stay at the top of the page when scrolling. After some thought I think keeping it at the top is better, so I put it back.