Added survey result pane to survey
If the option is turned on in the story form, participants can see a textarea with their survey questions and answers. They can copy from that area and paste it somewhere else to keep a local copy of what they submitted to the project. This is also so people who are entering stories collected elsewhere can keep a backup copy of each story as they enter it. It works both on the survey proper and on the Enter stories page. The default option is NOT to show this extra box on the survey form, so it has to be turned on in the story form to wok.